+++ react-native/React/Base/RCTKeyCommands.m
@@ -1,8 +1,10 @@
 /**
- * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
  *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
  */
 
 #import "RCTKeyCommands.h"
@@ -66,8 +68,8 @@ RCT_NOT_IMPLEMENTED(- (instancetype)init)
 
 - (NSString *)description
 {
-  return [NSString stringWithFormat:@"<%@:%p input=\"%@\" flags=%lld hasBlock=%@>",
-          [self class], self, _keyCommand.input, (long long)_keyCommand.modifierFlags,
+  return [NSString stringWithFormat:@"<%@:%p input=\"%@\" flags=%zd hasBlock=%@>",
+          [self class], self, _keyCommand.input, _keyCommand.modifierFlags,
           _block ? @"YES" : @"NO"];
 }
